+++ core/admin/login.php
@@ -9,8 +9,6 @@ $ps = filter_var(trim($_POST['ps']), FILTER_SANITIZE_STRING);
 if ((string) $ps === HOUND_PASS) {
     $_SESSION['temppass'] = $ps;
 
-    //houndLogAccess();
-
     php_redirect('dashboard.php');
 } else {
     php_redirect('index.php?err=1');